Fixed HexView Display Offset Bug
If there is data to display this line will write a hexadecimal byte as a string. If there is no data it will instead fill the gap with spaces.

A hexidecimal byte is represented by two digits (e.g., "9F"). Instead of filling 2 spaces to represent the byte it is filling in three spaces. This is causing the formatting of lines without data to be heavily offset and not lining up with lines that contain data.