Allow key events to traverse the TopLevel hierarchy unless IsModal=true (#262)
It's quite common for key events to flow up the containment hierarchy in
UI frameworks. This could simplify management of global hotkeys that can
be placed in a top-level "App" and have all views shown subsequently just
"inherit" that behavior as long as they let the key event flow.

Propagating keys upwards is typically undesirable in modal dialogs, so we
set it to true in the `Dialog` base class.

Make sure we make a copy of the items since a key handler could modify the
top-level collection (i.e. by closing a dialog in between this processing).