Refactor test cleanup and skip slow test
Removed redundant calls to `Application.Top?.Dispose()` and
`Application.ResetState()` from multiple test methods in the
`ColorPickerTests` class. This change likely reflects a shift
towards automatic or centralized test cleanup.

Marked the `MouseClick_ChangesTab` test in `TabViewTests` as
skipped by adding a `Skip` parameter to the `[Fact]` attribute
with the value `"Really slow test"`. This was done to address
performance concerns or temporarily disable the test.